+++ Particles/src/effect/Snow.java
@@ -14,7 +14,7 @@ public class Snow extends Effect {
     private static final short TOTAL_PARTICLES = 5000;
 
 	/** The length (x-axis) of the screen. */
-	private final double screenLength;
+	private final short screenLength;
 
     /**
      * Constructs a new Snow particle effect.
@@ -22,7 +22,7 @@ public class Snow extends Effect {
      * @param originY The origin, on the Y-axis, of the effect.
      * @param screenLength The total length of the screen.
      */
-	public Snow(final double originX, final double originY, final double screenLength) {
+	public Snow(final float originX, final float originY, final short screenLength) {
 		super(originX, originY - 50, true, TOTAL_PARTICLES);
 		this.screenLength = screenLength;
 	}
@@ -34,7 +34,7 @@ public class Snow extends Effect {
      * @param screenLength The total length of the screen.
      * @param totalParticles The total number of particles that this effect will use.
      */
-    public Snow(final double originX, final double originY, final double screenLength, final short totalParticles) {
+    public Snow(final float originX, final float originY, final short screenLength, final short totalParticles) {
         super(originX, originY - 50, true, totalParticles);
         this.screenLength = screenLength;
     }
@@ -67,14 +67,14 @@ public class Snow extends Effect {
         short indexOfOpenPosition = super.canTakeNewParticles();
 
         if(indexOfOpenPosition != -1) {
-            double xCoord = (int)(Math.random() * screenLength);
-            double yCoord = super.originY;
-            double dx = Math.random() * (Math.random() >= 0.5 ? -2 : 2);
-            double dy = Math.random() * 2.5;
-            double gravityX = -0.0075;
-            double gravityY = 0.0;
-            int size = (int)(Math.random() * 8 + 1);
-            double life = Math.random() * 800 + 1;
+            float xCoord = (float)(Math.random() * screenLength);
+            float yCoord = super.originY;
+            float dx = (float)Math.random() * (Math.random() >= 0.5f ? -2f : 2f);
+            float dy = (float)Math.random() * 2.5f;
+            float gravityX = -0.0075f;
+            float gravityY = 0.0f;
+            byte size = (byte)(Math.random() * 8 + 1);
+            short life = (short)(Math.random() * 800 + 1);
 
             super.addParticle(new Particle(xCoord, yCoord, dx, dy, gravityX, gravityY, size, life, Color.WHITE), indexOfOpenPosition);
         }
